All right, so we created our exercise API, recreated our exercise DbContext, and generated the seed data to populate the database to which the API is connected. It is a SQL Server database. After creating the seed data in the exercise DbContext class, we created a migration and updated the database. We also created an exercise model. Initially, we did not have a primary key (ID) in this model, which was causing errors. Ultimately, we added a primary key (ID), and we will need to create a DTO so that the ID doesn't appear in the JSON response. Lastly, we created an ExercisesController through the creation of a scaffolded item using actions and EF Core.
Next, we will need to create the DTO, inject the service, and install the NuGet package for AutoMapper. We also need to create a custom controller endpoint to get all exercises by workout type. In addition, we need to refactor the layout of this application to include a service for business logic, a repository to handle database operations, and our controller, which will manage the API requests and responses.